requirements.txt
bcrypt==5.0.0
cryptography==46.0.3
email-validator==2.3.0
fastapi==0.125.0
pydantic==2.12.5
pydantic_core==2.41.5
python-jose==3.5.0
uvicorn==0.38.0
